增加对二进制资源的支持

+++ GameFramework/Resource/DecryptResourceCallback.cs
@@ -18,5 +18,5 @@ namespace GameFramework.Resource
     /// <param name="storageInReadOnly">资源是否在只读区。</param>
     /// <param name="bytes">待解密的资源二进制流。</param>
     /// <returns>解密后的资源二进制流。</returns>
-    public delegate byte[] DecryptResourceCallback(string name, string variant, int loadType, int length, int hashCode, bool storageInReadOnly, byte[] bytes);
+    public delegate byte[] DecryptResourceCallback(string name, string variant, byte loadType, int length, int hashCode, bool storageInReadOnly, byte[] bytes);
 }
